Extract helper functions
