+++ playground/src/screens/sharedElementTransitionAlt/PostsListScreen.tsx
@@ -1,9 +1,10 @@
 import React, { useCallback } from 'react';
 import { Platform, SafeAreaView, ScrollView, StyleSheet, TouchableOpacity } from 'react-native';
 import { NavigationFunctionComponent } from 'react-native-navigation';
-import posts, { PostItem, SET_DURATION } from '../../assets/posts';
+import posts, { PostItem } from '../../assets/posts';
 import Navigation from '../../services/Navigation';
 import Screens from '../Screens';
+import { buildSharedElementAnimations } from './Constants';
 import PostCard from './PostCard';
 
 // SET = Shared Element Transition
@@ -18,60 +19,14 @@ import PostCard from './PostCard';
 
 const PostsListScreen: NavigationFunctionComponent = (props) => {
   const onPostPressed = useCallback(
-    (post: PostItem) => {
+    async (post: PostItem) => {
+      const navigationAnimations = await buildSharedElementAnimations(post);
       Navigation.push(props.componentId, {
         component: {
           name: Screens.PostDetailsScreen,
           passProps: { post: post },
           options: {
-            animations: {
-              push: {
-                content: {
-                  alpha: {
-                    from: 0,
-                    to: 1,
-                    duration: SET_DURATION,
-                  },
-                },
-                sharedElementTransitions: [
-                  {
-                    fromId: `image${post.id}`,
-                    toId: `image${post.id}Dest`,
-                    duration: SET_DURATION,
-                    interpolation: 'spring',
-                  },
-                  {
-                    fromId: `title${post.id}`,
-                    toId: `title${post.id}Dest`,
-                    duration: SET_DURATION,
-                    interpolation: 'spring',
-                  },
-                ],
-              },
-              pop: {
-                content: {
-                  alpha: {
-                    from: 1,
-                    to: 0,
-                    duration: SET_DURATION,
-                  },
-                },
-                sharedElementTransitions: [
-                  {
-                    fromId: `image${post.id}Dest`,
-                    toId: `image${post.id}`,
-                    duration: SET_DURATION,
-                    interpolation: 'spring',
-                  },
-                  {
-                    fromId: `title${post.id}Dest`,
-                    toId: `title${post.id}`,
-                    duration: SET_DURATION,
-                    interpolation: 'spring',
-                  },
-                ],
-              },
-            },
+            animations: navigationAnimations,
           },
         },
       });
